+++ sklearn/datasets/tests/test_svmlight_format.py
@@ -155,6 +155,23 @@ def test_load_zero_based_auto():
     assert_equal(X2.shape, (1, 4))
 
 
+def test_load_with_qid():
+    # load svmfile with qid attribute
+    data = """
+    3 qid:1 1:0.53 2:0.12
+    2 qid:1 1:0.13 2:0.1
+    7 qid:2 1:0.87 2:0.12"""
+    X, y = load_svmlight_file(BytesIO(data), query_id=False)
+    assert_array_equal(y, [3, 2, 7])
+    assert_array_equal(X.todense(), [[.53, .12], [.13, .1], [.87, .12]])
+    res1 = load_svmlight_files([BytesIO(data)], query_id=True)
+    res2 = load_svmlight_file(BytesIO(data), query_id=True)
+    for X, y, qid in (res1, res2):
+        assert_array_equal(y, [3, 2, 7])
+        assert_array_equal(qid, [1, 1, 2])
+        assert_array_equal(X.todense(), [[.53, .12], [.13, .1], [.87, .12]])
+
+
 @raises(ValueError)
 def test_load_invalid_file2():
     load_svmlight_files([datafile, invalidfile, datafile])
@@ -245,3 +262,18 @@ def test_dump_invalid():
 
     f = BytesIO()
     assert_raises(ValueError, dump_svmlight_file, X, y[:-1], f)
+
+
+def test_dump_query_id():
+    # test dumping a file with query_id
+    X, y = load_svmlight_file(datafile)
+    X = X.toarray()
+    query_id = np.arange(X.shape[0]) // 2
+    f = BytesIO()
+    dump_svmlight_file(X, y, f, query_id=query_id, zero_based=True)
+
+    f.seek(0)
+    X1, y1, query_id1 = load_svmlight_file(f, query_id=True, zero_based=True)
+    assert_array_almost_equal(X, X1.toarray())
+    assert_array_almost_equal(y, y1)
+    assert_array_almost_equal(query_id, query_id1)