+++ sklearn/pipeline.py
@@ -8,9 +8,13 @@ estimator, as a chain of transforms and estimators.
 #         Alexandre Gramfort
 # Licence: BSD
 
-from .base import BaseEstimator
+import numpy as np
+from scipy import sparse
 
-__all__ = ['Pipeline']
+from .base import BaseEstimator, TransformerMixin
+from .externals.joblib import Parallel, delayed
+
+__all__ = ['Pipeline', 'FeatureUnion']
 
 
 # One round of beers on me if someone finds out why the backslash
@@ -174,15 +178,15 @@ class Pipeline(BaseEstimator):
         final estimator. Valid only if the final estimator implements
         transform."""
         Xt = X
-        for name, transform in self.steps[:-1]:
+        for name, transform in self.steps:
             Xt = transform.transform(Xt)
-        return self.steps[-1][-1].transform(Xt)
+        return Xt
 
     def inverse_transform(self, X):
         if X.ndim == 1:
             X = X[None, :]
         Xt = X
-        for name, step in self.steps[:-1][::-1]:
+        for name, step in self.steps[::-1]:
             Xt = step.inverse_transform(Xt)
         return Xt
 
@@ -199,3 +203,102 @@ class Pipeline(BaseEstimator):
     def _pairwise(self):
         # check if first estimator expects pairwise input
         return getattr(self.steps[0][1], '_pairwise', False)
+
+
+def _fit_one_transformer(transformer, X, y):
+    transformer.fit(X, y)
+
+
+def _transform_one(transformer, name, X, transformer_weights):
+    if transformer_weights is not None and name in transformer_weights:
+        # if we have a weight for this transformer, muliply output
+        return transformer.transform(X) * transformer_weights[name]
+    return transformer.transform(X)
+
+
+class FeatureUnion(BaseEstimator, TransformerMixin):
+    """Concatenates results of multiple transformer objects.
+
+    This estimator applies a list of transformer objects in parallel to the
+    input data, then concatenates the results. This is useful to combine
+    several feature extraction mechanisms into a single transformer.
+
+    Parameters
+    ----------
+    transformers: list of (name, transformer)
+        List of transformer objects to be applied to the data.
+
+    n_jobs: int, optional
+        Number of jobs to run in parallel (default 1).
+
+    transformer_weights: dict, optional
+        Multiplicative weights for features per transformer.
+        Keys are transformer names, values the weights.
+
+    """
+    def __init__(self, transformer_list, n_jobs=1, transformer_weights=None):
+        self.transformer_list = transformer_list
+        self.n_jobs = n_jobs
+        self.transformer_weights = transformer_weights
+
+    def get_feature_names(self):
+        """Get feature names from all transformers.
+
+        Returns
+        -------
+        feature_names : list of strings
+            Names of the features produced by transform.
+        """
+        feature_names = []
+        for name, trans in self.transformer_list:
+            if not hasattr(trans, 'get_feature_names'):
+                raise AttributeError("Transformer %s does not provide"
+                        " get_feature_names." % str(name))
+            feature_names.extend([name + "__" + f
+                for f in trans.get_feature_names()])
+        return feature_names
+
+    def fit(self, X, y=None):
+        """Fit all transformers using X.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix, shape (n_samples, n_features)
+            Input data, used to fit transformers.
+        """
+        Parallel(n_jobs=self.n_jobs)(delayed(_fit_one_transformer)(trans, X, y)
+                for name, trans in self.transformer_list)
+        return self
+
+    def transform(self, X):
+        """Transform X separately by each transformer, concatenate results.
+
+        Parameters
+        ----------
+        X : array-like or sparse matrix, shape (n_samples, n_features)
+            Input data to be transformed.
+
+        Returns
+        -------
+        X_t : array-like or sparse matrix, shape (n_samples, sum_n_components)
+            hstack of results of transformers. sum_n_components is the
+            sum of n_components (output dimension) over transformers.
+        """
+        Xs = Parallel(n_jobs=self.n_jobs)(
+            delayed(_transform_one)(trans, name, X, self.transformer_weights)
+            for name, trans in self.transformer_list)
+        if any(sparse.issparse(f) for f in Xs):
+            Xs = sparse.hstack(Xs).tocsr()
+        else:
+            Xs = np.hstack(Xs)
+        return Xs
+
+    def get_params(self, deep=True):
+        if not deep:
+            return super(FeatureUnion, self).get_params(deep=False)
+        else:
+            out = dict(self.transformer_list)
+            for name, trans in self.transformer_list:
+                for key, value in trans.get_params(deep=True).iteritems():
+                    out['%s__%s' % (name, key)] = value
+            return out